Proj/GDMS/requirement.txt
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
decorator==4.4.1
Django==2.2
django-cors-headers==3.2.1
django-extensions==2.2.8
django-shell-plus==1.1.7
django-stubs==1.4.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.1.0
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
mysqlclient==1.4.6
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
wcwidth==0.1.8
wrapt==1.11.2
xlrd==1.2.0
